Re-organized testutils package to prevent overcrowding it
