+++ core/src/androidTest/java/com/amplifyframework/core/reachability/SocketHostInstrumentationTest.java
@@ -15,7 +15,7 @@
 
 package com.amplifyframework.core.reachability;
 
-import com.amplifyframework.testutils.RandomString;
+import com.amplifyframework.testutils.random.RandomString;
 
 import org.junit.Test;
