+++ rxbindings/src/test/java/com/amplifyframework/rx/RxHubBindingTest.java
@@ -17,7 +17,7 @@ package com.amplifyframework.rx;
 
 import com.amplifyframework.hub.HubChannel;
 import com.amplifyframework.hub.HubEvent;
-import com.amplifyframework.testutils.RandomString;
+import com.amplifyframework.testutils.random.RandomString;
 
 import org.junit.Before;
 import org.junit.Test;
