+++ testmodels/src/test/java/com/amplifyframework/testmodels/commentsblog/BlogOwnerTest.java
@@ -15,7 +15,7 @@
 
 package com.amplifyframework.testmodels.commentsblog;
 
-import com.amplifyframework.testutils.RandomString;
+import com.amplifyframework.testutils.random.RandomString;
 
 import org.junit.Test;
