+++ testutils/src/main/java/com/amplifyframework/testutils/random/RandomString.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.testutils;
+package com.amplifyframework.testutils.random;
 
 import java.util.UUID;
